restructure: experimental_devices files
